+++ packages/extensions/lib/hetu/hetu.dart
@@ -6,6 +6,7 @@ import './helpers/future.dart' as helpers;
 import './helpers/fuzzy.dart' as helpers;
 import './helpers/html.dart' as helpers;
 import './helpers/list.dart' as helpers;
+import './helpers/map.dart' as helpers;
 import './helpers/regexp.dart' as helpers;
 import './helpers/throw_err.dart' as helpers;
 import '../utils/http.dart';
@@ -36,6 +37,10 @@ Future<Hetu> createHetu() async {
       'resolveFuture': helpers.resolveFuture,
       'throwError': helpers.throwError,
       'ensureURL': HttpUtils.ensureURL,
+      'eachList': helpers.eachList,
+      'eachMap': helpers.eachMap,
+      'mergeMap': helpers.mergeMap,
+      'mergeList': helpers.mergeList,
     },
   );